Added DukeEnemyBullet to reflect the static damage of 5 that the Battlelord & Assault Enforcer do with each round.
Updated the Assault Enforcer and Battlelord classes accordingly.

Changed ACommanderRocket to better reflect the actual damage of the Duke3D counterpart.

Removed ZorchCrash from HereticImp3 as it no longer seems to exist in the latest version of Samsara.

Heightened the spawning chance for the Abaddon and Cacolantern from 19 to 199. (to prevent the SODBat from spawning as often)

Heightened the spawning chance for the Hectebus and Belphegor from 199 to 19,999. (to prevent the others from spawning as often)